rename and update script
